+++ exercises/mad_libs.livemd
@@ -38,7 +38,7 @@ secret_sauce = nil
 
 madlib = nil
 
-Utils.feedback(:startup_madlib_answers, [
+Utils.feedback(:startup_madlib, [
   madlib,
   name_of_company,
   a_defined_offering,
