+++ utils/test/feedback/mad_libs_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.MadLibsTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.MadLibs,
+    file: "../exercises/mad_libs.livemd"
+end
